change scale dtype

+++ sharktank/tests/kernels/elementwise_tensor_tensor_test.py
@@ -32,7 +32,7 @@ class elementwise_tensor_tensor_test(unittest.TestCase):
         shape = [17, 32]
         x = torch.rand(shape, dtype=dtype)
         y = torch.rand(shape, dtype=dtype)
-        one = torch.tensor(1.0).float()
+        one = torch.tensor(1.0).to(dtype=dtype)
         x_layout = TensorScaledLayout(shape=shape, qs=x, d=one)
         x_qtensor = PlanarQuantizedTensor(shape=shape, layout=x_layout)
         y_layout = TensorScaledLayout(shape=shape, qs=y, d=one)
